Moved SafeJSONEncoder from templatetags to cms.utils
