+++ cms/templatetags/cms_js_tags.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 import warnings
 from classytags.core import Tag, Options
+from cms.utils import SafeJSONEncoder
 from cms.utils.compat import DJANGO_1_4
 from django import template
 from django.core.serializers.json import DjangoJSONEncoder
-from django.utils.html import conditional_escape
 from django.utils.text import javascript_quote
 from django.utils.safestring import mark_safe
 if DJANGO_1_4:
@@ -19,22 +19,6 @@ def js(value):
     warnings.warn("The template filter '...|js' is vulnerable to XSS attacks, please use '...|json' instead.",
                   DeprecationWarning, stacklevel=2)
     return json.dumps(value, cls=DjangoJSONEncoder)
-
-
-class SafeJSONEncoder(DjangoJSONEncoder):
-    def _recursive_escape(self, o, esc=conditional_escape):
-        if isinstance(o, dict):
-            return type(o)((esc(k), self._recursive_escape(v)) for (k, v) in o.iteritems())
-        if isinstance(o, (list, tuple)):
-            return type(o)(self._recursive_escape(v) for v in o)
-        try:
-            return type(o)(esc(o))
-        except ValueError:
-            return esc(o)
-
-    def encode(self, o):
-        value = self._recursive_escape(o)
-        return super(SafeJSONEncoder, self).encode(value)
 
 
 @register.filter('json')
